Test array-api-tests and add ones_like (#23)
* Test array-api-tests and add ones_like

* ruff

* requirements

* */

* improvments

* better processing of tuple

* improvments

* add missing argument

* fix one issue

* refactoring

* documentation

* more coverage for the functions

* refactoring

* documentation

* fix issue with garbage collector and OrtTensor initilized from numpy array

* add a unit test for bfloat16

* disable one test for ones_like

* lint

* fix skipif

* api

+++ _unittests/ut_array_api/test_array_apis.py
@@ -18,7 +18,7 @@ class TestArraysApis(ExtTestCase):
     def test_zeros_ort_1(self):
         c = xpo.zeros(1)
         d = c.numpy()
-        self.assertEqualArray(np.array([0], dtype=np.float32), d)
+        self.assertEqualArray(np.array([0], dtype=np.float64), d)
 
     def test_ffinfo(self):
         dt = np.float32
